+++ Entidades/PrestamosDetalle.cs
@@ -13,7 +13,7 @@ namespace PrestamosdeJuegos.Entidades
         public int PrestamoDetalleId { get; set; }
         public int PrestamosId { get; set; }
         public int JuegoId { get; set; }
-        public int CantidadJuegos { get; set; }
+        public double CantidadJuegos { get; set; }
 
         [ForeignKey("JuegoId")]
         public virtual Juegos FK_Juegos { get; set; }
